+++ src/App.css
@@ -269,3 +269,21 @@ input:focus::placeholder {
 .animate-vibrate {
   animation: vibrate 0.69s linear;
 }
+
+@keyframes faint-right {
+  0% {
+    transform: none;
+    opacity: 1;
+  }
+  60% {
+    transform: rotate(90deg) translateX(40px) scale(1);
+    opacity: 0.7;
+  }
+  100% {
+    transform: rotate(90deg) translateX(120px) scale(0.7);
+    opacity: 0.3;
+  }
+}
+.animate-faint-right {
+  animation: faint-right 0.7s forwards;
+}
